I'm not sure why the fedpkg piece won't simply commit changes.
I made changes to .gitignore and sources but "git commit -a" *WON'T*
let me put them in as a single commit, which is ridiculous. So I have to
now add a change to the specfile, which I'm going to then remove later.
What a ridiculous thing to force a package maintainer to do when they
simply added a new source.